COMP: Use SuperBuild to download/configure/build dependencies
This commit updates the build-system to

(1) download, configure and build dependencies (LIBTIFF & GLEW)
    as part of an outer build.

(2) configure and build Autoscoper itself (the inner build)
    setting options and passing paths associated with the
    external dependencies.

Both inner and outer build are described using the top-level
CMakeLists.txt.

By default, the Autoscoper_SUPERBUILD option is ON and this
corresponds to the outer build and means that the
SuperBuild/External_*.cmake files are processed.

Once the external dependencies described in each External_*.cmake
files are processed, the last external project consists in
building the Autoscoper project itself with Autoscoper_SUPERBUILD
set to ON.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>